requirements.txt
stramlit>=1.29 #frontend Framework
supabase>=2.0.2 #supabase client for  database operations 
fastapi>=0.104.1 #backenk API framework
uvicorn>=0.24.0 #ASGI server for FastAPI
python-dotenv>=1.0.0 #Envirnonmental variables management